Add files via upload
New Patch - Full Human vs Human game functionality implemented. Full username validation is also implemented, and two valid usernames are required for the game to start. Game keeps track of how many times each player wins and resets this score along with the usernames when the game is restarted.

The functionality for the human to play against the computer, or for computer vs computer has not been implemented at this point. This will be implemented at a later stage, and is the last feature to be added.